+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+...
+
+## [0.2.0] - 2021-04-05
+
 ### Added
 - Reading the measurement of all channels at once.
 - Make types copyable, clonable, comparable and debuggable where possible.
@@ -37,6 +41,7 @@ This crate is now functionally complete.
 This is the initial release to crates.io. All changes will be documented in
 this CHANGELOG.
 
-[Unreleased]: https://github.com/eldruin/tcs3472-rs/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/eldruin/tcs3472-rs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/eldruin/tcs3472-rs/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/eldruin/tcs3472-rs/compare/v0.1.0...v0.1.1